Update: Soft Delete Implementation

+++ src/main/resources/db/migration/V2__create-table-body_weight.sql
@@ -1,7 +1,7 @@
 CREATE TABLE body_weight(
     id VARCHAR(255) NOT NULL PRIMARY KEY UNIQUE,
     user_id VARCHAR(255) NOT NULL,
-    value int NOT NULL,
+    value DOUBLE NOT NULL,
     is_active BOOLEAN NOT NULL,
     delete_on DATETIME,
     logged_on DATETIME NOT NULL,
